% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_five.R
\name{add_five}
\alias{add_five}
\title{Add five to numeric vector}
\usage{
add_five(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Add five to numeric vector
}
\examples{
add_five(2)
add_five(3)
}
