% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_four.R
\name{add_four}
\alias{add_four}
\title{Add four to numeric vector}
\usage{
add_four(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Add four to numeric vector
}
\examples{
add_four(2)
add_four(3)
}
