% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_six.R
\name{add_six}
\alias{add_six}
\title{Add six to numeric vector}
\usage{
add_six(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector
}
\description{
Add six to numeric vector
}
\examples{
add_six(2)
add_six(3)
}
